feet:Passed the function through the Vulkan API, now it works for everyone

+++ hardware.h
@@ -35,7 +35,7 @@ typedef struct {
     char* os_name;
 } Config;
 
-unsigned long long GetVRAM();
+int GetVRAM();
 GpuVendor GetGpuVendorByPciId();
 char* GetGpuVendorName(GpuVendor);
 char* GetGpuName();
